Refactored ext_modules to intmodules package
