--- ext_modules/__init__.py
@@ -1,4 +1,0 @@
-from ext_modules.log_module import config_logger
-from ext_modules.log_module import logger
-from ext_modules.preprocessor import preprocess_data
-from ext_modules.json_validation import validate_request
