+++ intmodules/preprocessor.py
@@ -1,5 +1,5 @@
-from ext_modules.reader import base64_to_dataframe
-from ext_modules.validator import validate_data
+from intmodules.reader import base64_to_dataframe
+from intmodules.validator import validate_data
 
 
 def preprocess_data(b64_data):
