+++ services/fitter/fitter.py
@@ -1,4 +1,4 @@
-from ext_modules import logger
+from intmodules import logger
 import numpy as np
 import matplotlib.pyplot as plt
 from scipy.optimize import curve_fit
